lambda-functions/updateUserProfile/requirements.txt
PyJWT==2.8.0
boto3==1.34.0
pydantic==2.12.4
boto3-stubs[dynamodb]==1.40.74